tests: Restructure file layout
Move mocks and resource factory related files to separate folder.
