+++ tests/mocks/dxgi_mocks.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include "nvapi_tests_private.h"
-#include "../src/interfaces/dxvk_interfaces.h"
-#include "../src/nvapi/nvml.h"
+#include "../../src/nvapi_private.h"
+#include "../../src/interfaces/dxvk_interfaces.h"
 
 class IDXGIDxvkFactoryMock : public IDXGIFactory1, public IDXGIVkInteropFactory1 {};
 
@@ -80,23 +79,4 @@ class DXGIOutput6Mock final : public trompeloeil::mock_interface<IDXGIOutput6> {
     IMPLEMENT_MOCK5(DuplicateOutput1);
     IMPLEMENT_MOCK1(GetDesc1);
     IMPLEMENT_MOCK1(CheckHardwareCompositionSupport);
-};
-
-class NvmlMock final : public trompeloeil::mock_interface<dxvk::Nvml> {
-    IMPLEMENT_CONST_MOCK0(IsAvailable);
-    IMPLEMENT_CONST_MOCK1(ErrorString);
-    IMPLEMENT_CONST_MOCK2(DeviceGetHandleByPciBusId_v2);
-    IMPLEMENT_CONST_MOCK2(DeviceGetMemoryInfo_v2);
-    IMPLEMENT_CONST_MOCK2(DeviceGetPciInfo_v3);
-    IMPLEMENT_CONST_MOCK3(DeviceGetClockInfo);
-    IMPLEMENT_CONST_MOCK3(DeviceGetTemperature);
-    IMPLEMENT_CONST_MOCK3(DeviceGetThermalSettings);
-    IMPLEMENT_CONST_MOCK2(DeviceGetPerformanceState);
-    IMPLEMENT_CONST_MOCK2(DeviceGetUtilizationRates);
-    IMPLEMENT_CONST_MOCK3(DeviceGetVbiosVersion);
-    IMPLEMENT_CONST_MOCK2(DeviceGetCurrPcieLinkWidth);
-    IMPLEMENT_CONST_MOCK2(DeviceGetIrqNum);
-    IMPLEMENT_CONST_MOCK2(DeviceGetNumGpuCores);
-    IMPLEMENT_CONST_MOCK2(DeviceGetBusType);
-    IMPLEMENT_CONST_MOCK2(DeviceGetDynamicPstatesInfo);
 };
