+++ tests/nvapi/mock_factory.h
@@ -1,10 +1,11 @@
 #pragma once
 
 #include "nvapi_tests_private.h"
-#include "../src/nvapi/nvapi_resource_factory.h"
-#include "nvapi_d3d_mocks.h"
-#include "nvapi_sysinfo_mocks.h"
-#include "nvapi_vulkan_mocks.h"
+#include "../../src/nvapi/nvapi_resource_factory.h"
+#include "../mocks/vulkan_mocks.h"
+#include "../mocks/dxgi_mocks.h"
+#include "../mocks/nvml_mocks.h"
+#include "../mocks/lfx_mocks.h"
 
 using namespace trompeloeil;
