+++ tests/nvapi/resource_factory_util.cpp
