+++ tests/nvapi_d3d.cpp
@@ -1,6 +1,6 @@
 #include "nvapi_tests_private.h"
-#include "resource_factory_util.h"
-#include "nvapi_d3d_mocks.h"
+#include "mocks/d3d_mocks.h"
+#include "nvapi/resource_factory_util.h"
 
 using namespace trompeloeil;
