+++ tests/nvapi_d3d12.cpp
@@ -1,10 +1,8 @@
 #include "nvapi_tests_private.h"
-#include "resource_factory_util.h"
-#include "nvapi_sysinfo_mocks.h"
-#include "nvapi_vulkan_mocks.h"
-#include "nvapi_d3d12_mocks.h"
-#include "nvapi_d3d11_mocks.h"
-#include "nvapi_d3d_mocks.h"
+#include "mocks/d3d_mocks.h"
+#include "mocks/d3d11_mocks.h"
+#include "mocks/d3d12_mocks.h"
+#include "nvapi/resource_factory_util.h"
 
 using namespace trompeloeil;
