+++ tests/nvapi_main.cpp
@@ -1,6 +1,6 @@
 #define CATCH_CONFIG_MAIN
 #include "../inc/catch_amalgamated.hpp"
-#include "section_listener.h"
+#include "nvapi/section_listener.h"
 
 CATCH_REGISTER_TAG_ALIAS("[@unit-tests]", "[d3d],[d3d11],[d3d12],[drs],[sysinfo],[sysinfo-topo],[sysinfo-nvml],[sysinfo-hdr],[util],[vulkan]")
 CATCH_REGISTER_TAG_ALIAS("[@system]", "[system]")
