+++ tests/nvapi_sysinfo_nvml.cpp
@@ -1,7 +1,5 @@
 #include "nvapi_tests_private.h"
-#include "resource_factory_util.h"
-#include "nvapi_sysinfo_mocks.h"
-#include "nvapi_vulkan_mocks.h"
+#include "nvapi/resource_factory_util.h"
 
 using namespace trompeloeil;
 using namespace Catch::Matchers;
