+++ tests/nvapi_tests_private.h
@@ -10,5 +10,6 @@
 #include "../src/nvapi/nvapi_d3d_low_latency_device.h"
 #include "../src/nvapi/nvapi_d3d11_device.h"
 #include "../src/nvapi/nvapi_d3d12_device.h"
+
 #include "../inc/catch_amalgamated.hpp"
 #include "../inc/catch2/trompeloeil.hpp"
