+++ tests/nvapi_vulkan.cpp
@@ -1,6 +1,5 @@
 #include "nvapi_tests_private.h"
-#include "resource_factory_util.h"
-#include "nvapi_vulkan_mocks.h"
+#include "nvapi/resource_factory_util.h"
 
 using namespace trompeloeil;
